+++ inc/wp-widgets/class-widget-social-icons.php
@@ -68,6 +68,7 @@ class Wolf_Core_Widget_Socials extends WP_Widget {
 				'wolf_core_widget_socials_args',
 				array(
 					'services' => $socials,
+					'alignment'=> 'left',
 				)
 			)
 		);